merging PR #343: More browser support for voters.html
https://github.com/chaosbot/Chaos/pull/343: More browser support for voters.html

Description:
Before, this code used Object.values, which is only supported in Chrome and Firefox. This new code works replaces the line by mapping the keys array to the variable with all the data, to get the values for each key, causing the same effect. This is supported in almost all modern browsers, and IE9+. 

:ok_woman: PR passed with a vote of 8 for and 0 against, with a weighted total             of 7.5 and a threshold of 6.6.

Vote record:
@DrOfAwesomeness: 1
@MUCHZER: 1
@Smittyvb: 1
@SylvainThrd: 1
@amoffat: 1
@connorhartley: 1
@eukaryote31: 1
@phil-r: 1