merging PR #350: Send chaosbot commands via issue comments
https://github.com/chaosbot/Chaos/pull/350: Send chaosbot commands via issue comments

Description:
Alright I'm not sure what happened with the other branch, but here's a clean version.

There's a lot going on, so a brief explanation is as follows.

This provides the functionality to send commands to chaosbot by commenting with a command (see below)  on any given issue. Similar to how voting for PRs goes, there's a window in which votes are gathered. As it is currently set, any negative vote will cause the command to not be run.

Current commands include:
- [x] reopening issues
- [x] closing issues

Suggestions for new commands are welcome and can be implemented in the future.

Thanks go out to @phil-r  for providing a clean version.

Reopening this since it had full support last time, and just missed 1 vote. Tagging WIP until it reaches a passing state

:ok_woman: PR passed with a vote of 11 for and 0 against, with a weighted total             of 11.0 and a threshold of 6.7.

Vote record:
@Ealhad: 1
@MUCHZER: 1
@PlasmaPower: 1
@SylvainThrd: 1
@jackmawer: 1
@mark-i-m: 1
@md678685: 1
@phil-r: 1
@qgustavor: 1
@rhengles: 1
@rudehn: 1