merging PR #497: Implementing the DB stuff
https://github.com/chaosbot/Chaos/pull/497: Implementing the DB stuff

Description:
This introduces the database into chaosbot. Currently on my fork, the issue commands and meritocracy tables are implemented and working, and those are provided with this PR.

The API for the database ORM is pretty simple. See [here](http://docs.peewee-orm.com/en/latest/peewee/quickstart.html) for a quickstart to doing various operations.

Hope you all enjoy! This took a surprising amount of work to get implemented.

:white_check_mark: PR passed with a vote of 11 for and 1 against, a weighted total of 10.0 and a threshold of 6.5, and a current meritocracy review.

Vote record:
@Leigende: 1
@PlasmaPower: 1
@Swizz: 1
@andrewda: 1
@anythingbot: -1
@e-beach: 1
@ike709: 1
@justync7: 1
@mark-i-m: 1
@md678685: 1
@rhengles: 1
@rudehn: 1