merging PR #523: HOTFIX - vote spam
https://github.com/chaosbot/Chaos/pull/523: HOTFIX - vote spam

Description:
This is a hotfix to help stop the vote spam we are seeing. Here's how it works.

1. A command is identified and ran
2. The command is then put into the InactiveCommands table to signify it ran

The previous problem was that when identifying new commands, we didn't check if it was already in the inactive commands list.

What does this PR solve? Well it won't re-add commands to the db to be run. HOWEVER it won't solve the problem if there is an issue removing commands from the ActiveCommand table.

Either way, this check needs to be in.

Also, my apologies for the goof. I thought I had tested pretty well but something must have slipped through the cracks.

:white_check_mark: PR passed with a vote of 11 for and 0 against, a weighted total of 10.5 and a threshold of 6.2, and a current meritocracy review.

Vote record:
@Ajedi32: 1
@Leigende: 1
@Reuh: 1
@Smittyvb: 1
@Swizz: 1
@TimvdLippe: 1
@andrewda: 1
@paris-ci: 1
@qgustavor: 1
@rhengles: 1
@rudehn: 1